quillpost/comments/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
mysqlclient==2.2.0
django-mysql==4.12.0
requests==2.31.0






